feat: refactor EditExercise component to simplify exercise data handling and improve UI responsiveness
