+++ frontend/src/components/Exercise/EditExercise.vue
@@ -5,9 +5,7 @@
       :title="
         isViewExercise
           ? 'Exercise'
-          : isViewWorkoutExercise
-            ? 'Edit Exercise in workout'
-            : 'Edit Exercise'
+          : 'Edit Exercise'
       "
       @close="emit('close')"
     >
@@ -31,27 +29,22 @@
       <div class="mx-5">
         <div class="py-4">
           <h1 class="text-h5 font-weight-bold">
-            {{
-              isWorkoutExercise(selectedExercise)
-                ? selectedExercise.exercise.name
-                : selectedExercise.name
-            }}
+            {{ selectedExercise.name }}
           </h1>
           <p>
-            {{
-              isWorkoutExercise(selectedExercise)
-                ? selectedExercise.exercise.description
-                : selectedExercise.description
-            }}
+            {{ selectedExercise.description }}
           </p>
-          <div class="d-flex ga-2 align-center mt-2 flex-wrap">
+          <div
+            v-if="props.selectedExercise?.muscleGroups"
+            class="d-flex ga-2 align-center mt-2 flex-wrap"
+          >
             <v-chip
-              v-for="group in getMuscleGroupsForExercise()"
-              :key="group"
+              v-for="group in props.selectedExercise.muscleGroups"
+              :key="group.id"
               color="green-lighten-1"
               label
             >
-              {{ group }}
+              {{ group.name }}
             </v-chip>
           </div>
         </div>
@@ -67,11 +60,7 @@
               <v-list-item-title>
                 <span class="font-weight-medium">Default Sets:</span>
                 <span class="ml-2">
-                  {{
-                    isWorkoutExercise(selectedExercise)
-                      ? selectedExercise.sets
-                      : selectedExercise.defaultSets || 0
-                  }}
+                  {{ selectedExercise.defaultSets || 0 }}
                 </span>
               </v-list-item-title>
             </v-list-item>
@@ -84,11 +73,7 @@
               <v-list-item-title>
                 <span class="font-weight-medium">Default Reps:</span>
                 <span class="ml-2">
-                  {{
-                    isWorkoutExercise(selectedExercise)
-                      ? selectedExercise.reps
-                      : selectedExercise.defaultReps || 0
-                  }}
+                  {{ selectedExercise.defaultReps || 0 }}
                 </span>
               </v-list-item-title>
             </v-list-item>
@@ -101,11 +86,7 @@
               <v-list-item-title>
                 <span class="font-weight-medium">Default Pause:</span>
                 <span class="ml-2">
-                  {{
-                    isWorkoutExercise(selectedExercise)
-                      ? selectedExercise.pauseSeconds
-                      : selectedExercise.defaultPauseSeconds || 0
-                  }}
+                  {{ selectedExercise.defaultPauseSeconds || 0 }}
                   seconds
                 </span>
               </v-list-item-title>
@@ -121,13 +102,9 @@
                 <span class="font-weight-medium">Created at:</span>
                 <span class="ml-2">
                   {{
-                    isWorkoutExercise(selectedExercise)
-                      ? new Date(
-                        selectedExercise.exercise.createdAt,
-                      ).toLocaleDateString()
-                      : new Date(
-                        selectedExercise.createdAt,
-                      ).toLocaleDateString()
+                    new Date(
+                      selectedExercise.createdAt,
+                    ).toLocaleDateString()
                   }}
                 </span>
               </v-list-item-title>
@@ -142,13 +119,9 @@
                 <span class="font-weight-medium">Updated at:</span>
                 <span class="ml-2">
                   {{
-                    isWorkoutExercise(selectedExercise)
-                      ? new Date(
-                        selectedExercise.exercise.updatedAt,
-                      ).toLocaleDateString()
-                      : new Date(
-                        selectedExercise.updatedAt,
-                      ).toLocaleDateString()
+                    new Date(
+                      selectedExercise.updatedAt,
+                    ).toLocaleDateString()
                   }}
                 </span>
               </v-list-item-title>
@@ -169,29 +142,21 @@
             class="pt-2 d-flex ga-5 flex-column"
           >
             <v-text-field
-              v-model="editExercise.sets"
+              v-model="editExercise.defaultSets"
               label="Sets"
               type="number"
               variant="outlined"
               hide-details
             />
             <v-text-field
-              v-model="editExercise.reps"
+              v-model="editExercise.defaultReps"
               label="Reps"
               type="number"
               variant="outlined"
               hide-details
             />
             <v-text-field
-              v-if="isViewWorkoutExercise"
-              v-model="editExercise.weight"
-              label="Weight (kg)"
-              type="number"
-              variant="outlined"
-              hide-details
-            />
-            <v-text-field
-              v-model="editExercise.pauseSeconds"
+              v-model="editExercise.defaultPauseSeconds"
               label="Pause (seconds)"
               type="number"
               variant="outlined"
@@ -204,127 +169,54 @@
   </div>
 </template>
 <script setup lang="ts">
-import type { Exercise as workoutExercise } from '@/interfaces/Workout.interface';
-import type { Exercise } from '@/interfaces/Exercise.interface';
-import type { MuscleGroup } from '@/interfaces/MuscleGroup.interface';
-import type { AddExerciseToWorkout } from '@/interfaces/Workout.interface';
-import { useMuscleGroupStore } from '@/stores/muscleGroup.store';
+import type { Exercise, UpdateExercise } from '@/interfaces/Exercise.interface';
 import { useExerciseStore } from '@/stores/exercise.store';
-import {
-  updateExerciseInWorkout,
-  removeExercisesFromWorkout,
-} from '@/services/workout.service';
 import {
   updateExercise as updateExerciseInExercise,
   deleteExercise,
 } from '@/services/exercise.service';
-import { useWorkoutStore } from '@/stores/workout.store';
 import { toast } from 'vuetify-sonner';
 
 const props = defineProps<{
   workoutId?: number;
-  selectedExercise: workoutExercise | Exercise | null;
+  selectedExercise: Exercise | null;
   isViewExercise: boolean;
-  isViewWorkoutExercise: boolean;
 }>();
 
 const isViewExercise = ref(props.isViewExercise);
 
 const exerciseStore = useExerciseStore();
-const workoutStore = useWorkoutStore();
-const muscleGroupStore = useMuscleGroupStore();
 const isLoading = ref<boolean>(false);
 
-const editExercise = ref<AddExerciseToWorkout | null>({
-  exerciseId: isWorkoutExercise(props.selectedExercise)
-    ? Number(props.selectedExercise.exercise.id)
-    : Number(props.selectedExercise?.id || 0),
-  sets: isWorkoutExercise(props.selectedExercise)
-    ? props.selectedExercise.sets
-    : props.selectedExercise?.defaultSets || 0,
-  reps: isWorkoutExercise(props.selectedExercise)
-    ? props.selectedExercise.reps
-    : props.selectedExercise?.defaultReps || 0,
-  pauseSeconds: isWorkoutExercise(props.selectedExercise)
-    ? props.selectedExercise?.pauseSeconds
-    : props.selectedExercise?.defaultPauseSeconds || 0,
-  order: isWorkoutExercise(props.selectedExercise)
-    ? props.selectedExercise?.order
-    : 0,
-  weight: isWorkoutExercise(props.selectedExercise)
-    ? props.selectedExercise?.weight
-    : 0,
+const editExercise = ref<UpdateExercise | null>({
+  id:  Number(props.selectedExercise?.id || 0),
+  name: props.selectedExercise?.name || '',
+  description: props.selectedExercise?.description || '',
+  image: props.selectedExercise?.image || null,
+  muscleGroups: props.selectedExercise?.muscleGroups?.map((group) => group.id) || [],
+  defaultSets: props.selectedExercise?.defaultSets || 0,
+  defaultReps:props.selectedExercise?.defaultReps || 0,
+  defaultPauseSeconds:  props.selectedExercise?.defaultPauseSeconds || 0,
 });
-
-function isWorkoutExercise(
-  exercise: workoutExercise | Exercise | null,
-): exercise is workoutExercise {
-  return !!exercise && 'exercise' in exercise;
-}
 
 const emit = defineEmits<{
   (e: 'close'): void;
 }>();
 
-const getMuscleGroupsForExercise = (): string[] => {
-  if (!props.selectedExercise) {
-    return [];
-  }
-
-  const muscleGroup = muscleGroupStore.muscleGroups as MuscleGroup[];
-  if (isWorkoutExercise(props.selectedExercise)) {
-    return (props.selectedExercise.exercise.muscleGroups || []) 
-      .map(
-        (groupId) => muscleGroup.find((group) => group.id === +groupId)?.name,
-      )
-      .filter((name): name is string => !!name);
-  }
-
-  if (props.selectedExercise.muscleGroups) {
-    return (props.selectedExercise.muscleGroups || [])
-      .map(
-        (groupId) => muscleGroup.find((group) => group.id === +groupId)?.name,
-      )
-      .filter((name): name is string => !!name);
-  }
-
-  return [];
-};
-
 const removeExercise = async () => {
   try {
-    if (!props.selectedExercise || props.isViewWorkoutExercise && !props.workoutId) {
+    if (!props.selectedExercise && !props.workoutId) {
       console.error('No exercise or workout ID provided.');
       return;
     }
 
     let response = null;
-    if (props.isViewWorkoutExercise) {
-      if (!props.workoutId) {
-        console.error('No workout ID provided for removing exercise from workout.');
-        return;
-      }
-      response = await removeExercisesFromWorkout(
-        props.workoutId,
-        isWorkoutExercise(props.selectedExercise)
-          ? [props.selectedExercise.exercise.id]
-          : [],
-      );
-    } else {
-      response = await deleteExercise(
-        isWorkoutExercise(props.selectedExercise)
-          ? props.selectedExercise.exercise.id
-          : props.selectedExercise?.id ?? 0,
-            );
-    }
 
+    response = await deleteExercise(props.selectedExercise?.id ?? 0);
+    
     if (response) {
       toast.success('Exercise removed successfully!', { progressBar: true });
-      if (props.isViewWorkoutExercise) {
-        await workoutStore.setWorkouts(true);
-      } else {
         await exerciseStore.setExercises(true);
-      }
       emit('close');
     } else {
       console.error('Failed to remove exercise.');
@@ -334,100 +226,18 @@ const removeExercise = async () => {
   }
 };
 
-const getSanitizedExerciseDataForWorkout = () => {
-  if (!props.selectedExercise || !editExercise.value) return {};
-
-  const original = isWorkoutExercise(props.selectedExercise)
-    ? {
-        sets: props.selectedExercise.sets,
-        reps: props.selectedExercise.reps,
-        pauseSeconds: props.selectedExercise.pauseSeconds,
-        order: props.selectedExercise.order,
-        weight: props.selectedExercise.weight,
-        exerciseId: Number(props.selectedExercise.exercise.id),
-      }
-    : {};
-
-  const edited = {
-    sets: Number(editExercise.value.sets || 0),
-    reps: Number(editExercise.value.reps || 0),
-    pauseSeconds: Number(editExercise.value.pauseSeconds || 0),
-    order: Number(editExercise.value.order || 0),
-    weight: Number(editExercise.value.weight || 0),
-    exerciseId: Number(editExercise.value.exerciseId || 0),
-  };
-
-  return Object.fromEntries(
-    (Object.entries(edited) as [keyof typeof edited, number][]).filter(
-      ([key, value]) => (original as Record<string, number>)[key as string] !== value
-    )
-  );
-};
-
 const getSanitizedExerciseData = () => {
   return {
-    name: isWorkoutExercise(props.selectedExercise)
-      ? props.selectedExercise.exercise.name
-      : props.selectedExercise?.name || '',
-    description: isWorkoutExercise(props.selectedExercise)
-      ? props.selectedExercise.exercise.description
-      : props.selectedExercise?.description || '',
-    defaultSets: Number(editExercise.value?.sets || 0),
-    defaultReps: Number(editExercise.value?.reps || 0),
-    defaultPauseSeconds: Number(editExercise.value?.pauseSeconds || 0),
-    muscleGroupIds: isWorkoutExercise(props.selectedExercise)
-      ? props.selectedExercise.exercise.muscleGroups?.map(Number) || []
-      : props.selectedExercise?.muscleGroups?.map(Number) || [],
+    name: props.selectedExercise?.name || '',
+    description: props.selectedExercise?.description || '',
+    defaultSets: Number(editExercise.value?.defaultSets || 0),
+    defaultReps: Number(editExercise.value?.defaultReps || 0),
+    defaultPauseSeconds: Number(editExercise.value?.defaultPauseSeconds || 0),
+    muscleGroupIds: props.selectedExercise?.muscleGroups?.map(Number) || [],
   };
 };
 
 const updateExercise = async () => {
-  if (props.isViewWorkoutExercise) {
-    try {
-      isLoading.value = true;
-      if (!editExercise.value) {
-        toast.error('No exercise data to update.');
-        return;
-      }
-      if (!props.workoutId) {
-        toast.error('No workout ID provided.');
-        return;
-      }
-      if (!isWorkoutExercise(props.selectedExercise)) {
-        toast.error('Selected exercise is not a workout exercise.');
-        return;
-      }
-      const workoutExercise = workoutStore.currentWorkout?.exercises.find(
-        (ex) =>
-          isWorkoutExercise(props.selectedExercise)
-            ? ex.id === props.selectedExercise.id
-            : ex.exercise.id === props.selectedExercise?.id
-      );
-
-      if (!workoutExercise) {
-        toast.error('Could not find the exercise in the workout.');
-        return;
-      }
-
-      const response = await updateExerciseInWorkout(
-        props.workoutId,
-        workoutExercise.id,
-        getSanitizedExerciseDataForWorkout() || {},
-      );
-      if (response) {
-        toast.success('Exercise updated successfully!', { progressBar: true });
-        await workoutStore.setWorkouts(true);
-        emit('close');
-      } else {
-        toast.error('Failed to update exercise.');
-      }
-    } catch (error) {
-      toast.error('Error in updateExercise.');
-      console.error('Error in updateExercise:', error);
-    } finally {
-      isLoading.value = false;
-    }
-  } else {
     try {
       isLoading.value = true;
       if (!editExercise.value) {
@@ -439,9 +249,7 @@ const updateExercise = async () => {
         return;
       }
       const response = await updateExerciseInExercise(
-        isWorkoutExercise(props.selectedExercise)
-          ? props.selectedExercise.exercise.id
-          : props.selectedExercise?.id,
+        props.selectedExercise?.id,
         getSanitizedExerciseData() || {},
       );
       if (response) {
@@ -458,5 +266,4 @@ const updateExercise = async () => {
       isLoading.value = false;
     }
   }
-};
 </script>
